Quick commit of temperature alarm.
I wrote this file this morning as a quick way to prevent myself from playing games until my laptop overheats and powers off. It's unlikely that anyone else will be interested in it (though if you are, feel free to use it!), but I wanted some sort of record of it.